requirements_mac.txt
# PatchCore Educational Dashboard Requirements
# Core dashboard dependencies
dash>=3.2.0
plotly>=5.17.0
numpy>=1.24.0
Pillow>=10.0.0
dash-bootstrap-components>=1.4.0

# Jupyter notebook dependencies  
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# PatchCore & Anomaly Detection dependencies (optional for advanced tutorials)
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0

# Versione CPU di FAISS per Mac
faiss-cpu

# Optional: Full anomalib stack for advanced tutorials
# Uncomment the following lines when ready for production anomaly detection:
# anomalib[full]>=1.0.0
# opencv-python>=4.8.0
# albumentations>=1.3.0
